22w06a

+++ minecraft/src/net/minecraft/client/MouseHandler.java
@@ -124,17 +124,17 @@ public class MouseHandler {
 					}
 
 					this.accumulatedScroll += f;
-					float i = (float)((int)this.accumulatedScroll);
-					if (i == 0.0F) {
+					int i = (int)this.accumulatedScroll;
+					if (i == 0) {
 						return;
 					}
 
 					this.accumulatedScroll -= (double)i;
 					if (this.minecraft.player.isSpectator()) {
 						if (this.minecraft.gui.getSpectatorGui().isMenuActive()) {
-							this.minecraft.gui.getSpectatorGui().onMouseScrolled((double)(-i));
+							this.minecraft.gui.getSpectatorGui().onMouseScrolled(-i);
 						} else {
-							float j = Mth.clamp(this.minecraft.player.getAbilities().getFlyingSpeed() + i * 0.005F, 0.0F, 0.2F);
+							float j = Mth.clamp(this.minecraft.player.getAbilities().getFlyingSpeed() + (float)i * 0.005F, 0.0F, 0.2F);
 							this.minecraft.player.getAbilities().setFlyingSpeed(j);
 						}
 					} else {
