+++ minecraft/src/net/minecraft/client/player/RemotePlayer.java
@@ -54,7 +54,7 @@ public class RemotePlayer extends AbstractClientPlayer {
 		}
 
 		if (this.lerpHeadSteps > 0) {
-			this.yHeadRot = (float)((double)this.yHeadRot + Mth.wrapDegrees(this.lyHeadRot - (double)this.yHeadRot) / (double)this.lerpHeadSteps);
+			this.yHeadRot = this.yHeadRot + (float)(Mth.wrapDegrees(this.lyHeadRot - (double)this.yHeadRot) / (double)this.lerpHeadSteps);
 			this.lerpHeadSteps--;
 		}
