+++ minecraft/src/net/minecraft/world/entity/monster/Zombie.java
@@ -165,7 +165,7 @@ public class Zombie extends Monster {
 	@Override
 	protected int getExperienceReward(Player player) {
 		if (this.isBaby()) {
-			this.xpReward = (int)((float)this.xpReward * 2.5F);
+			this.xpReward = (int)((double)this.xpReward * 2.5);
 		}
 
 		return super.getExperienceReward(player);
